# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Raspored",
    "version": "v1"
  },
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserProfileDTO": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "yearOfEmployment": {
            "type": "integer",
            "format": "int32"
          },
          "licenseNumber": {
            "type": "string",
            "nullable": true
          },
          "contractTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "positionId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RegistrationDTO": {
        "required": [
          "dateOfBirth",
          "email",
          "firstName",
          "lastName",
          "licenseNumber",
          "password",
          "userName",
          "yearOfEmployment"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "userName": {
            "maxLength": 60,
            "minLength": 3,
            "pattern": "^(?=.{3,32}$)(?![_.-])(?!.*[_.]{2})[a-zA-Z0-9._-]+(?<![_.])$",
            "type": "string"
          },
          "email": {
            "pattern": "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$",
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 20,
            "minLength": 6,
            "pattern": "^(?=(?:.*[A-Z]){1,})(?=(?:.*[a-z]){1,})(?=(?:.*\\d){1,})(?=(?:.*[!@#$%^&*()\\-_=+{};:,<.>]){1,})(?!.*(.)\\1{})([A-Za-z0-9!@#$%^&*()\\-_=+{};:,<.>]{6,20})$",
            "type": "string"
          },
          "yearOfEmployment": {
            "maximum": 2023,
            "minimum": 1975,
            "type": "integer",
            "format": "int32"
          },
          "licenseNumber": {
            "type": "string"
          },
          "contractTypeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "positionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "registrationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LoginDTO": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}